modify VerticalBar to treat a bar of height 1 as a special case, namely don't insert a newline after it.
reasoning: because the vertical bar in its previous form expected that it will end on a new line, it used `Console.WriteLine`. The exception case was introduced because vertical bar required that the carret wasn't advanced to a new line, since the white space would follow between the bars. If the case wouldn't exist, there would be alignment issues that would destroy the shape of the rectangle. In that case, if the height is 1, only `Console.Write` is used.